LPF-190 - test that talks to endpoint, cleanup
