+++ src/test/java/uk/gov/laa/pfla/ServiceManager.java
@@ -8,7 +8,7 @@ public class ServiceManager {
 
     public void startService() throws IOException {
         String jarPath = new File("target/libs/pay-for-legal-aid-0.0.1-SNAPSHOT.jar").getAbsolutePath();
-        ProcessBuilder processBuilder = new ProcessBuilder( "java", "-jar", "-Dspring.profiles.active=local", jarPath);
+        ProcessBuilder processBuilder = new ProcessBuilder("java", "-jar", "-Dspring.profiles.active=local", jarPath);
         processBuilder.redirectOutput(ProcessBuilder.Redirect.INHERIT);
         processBuilder.redirectError(ProcessBuilder.Redirect.INHERIT);
         serviceProcess = processBuilder.start();
@@ -19,4 +19,5 @@ public class ServiceManager {
             serviceProcess.destroy();
         }
     }
+
 }
